refactor: P1-B 边界清理 — 移除 core/agent 的 AI 类型 re-export
Core 和 Agent 不再 re-export @zhin.js/ai 的类型，由消费方直接 import。
zhin.js（用户入口）保留常用 AI 类型的 re-export 作为便利导出。

变更：
- packages/im/core/src/index.ts: 移除 40 行 AI 类型/值 re-export
- packages/im/agent/src/index.ts: 移除 70 行 AI 类型/值 re-export
  （70 行 AI 代理 → 0 行，Agent 只导出自身模块）
- packages/im/zhin/src/index.ts: 将 AI 类型从 @zhin.js/agent 移至 @zhin.js/ai
- 修复 28 个 agent 文件的 import 路径（@zhin.js/core → @zhin.js/ai）
- 修复 provider-instance.ts 和 service.ts 的残留引用

效果: -110 行过度 re-export，模块边界更清晰
