+++ packages/im/agent/src/index.ts
@@ -5,78 +5,6 @@
  * tools, skills, subagents, mcps, and hooks.
  */
 
-// ── Re-export AI primitives from @zhin.js/ai (no shims) ──
-export type {
-  AIConfig,
-  AIProvider,
-  ChatMessage,
-  ContentPart,
-  AgentTool,
-  AgentConfig,
-  AgentResult,
-  ToolResultTransformInput,
-  ToolResultTransform,
-} from '@zhin.js/ai';
-
-export {
-  SessionManager,
-  MemorySessionManager,
-  DatabaseSessionManager,
-  createMemorySessionManager,
-  createDatabaseSessionManager,
-  AI_SESSION_MODEL,
-} from '@zhin.js/ai';
-
-export {
-  ContextManager,
-  createContextManager,
-  CHAT_MESSAGE_MODEL,
-  CONTEXT_SUMMARY_MODEL,
-} from '@zhin.js/ai';
-export type { ContextConfig, MessageRecord } from '@zhin.js/ai';
-
-export {
-  ConversationMemory,
-  AI_MESSAGE_MODEL,
-  AI_SUMMARY_MODEL,
-} from '@zhin.js/ai';
-export type { ConversationMemoryConfig } from '@zhin.js/ai';
-
-export { RateLimiter } from '@zhin.js/ai';
-export type { RateLimitConfig, RateLimitResult } from '@zhin.js/ai';
-
-export { detectTone } from '@zhin.js/ai';
-export type { Tone } from '@zhin.js/ai';
-
-export { parseOutput, renderToPlainText, renderToSatori } from '@zhin.js/ai';
-export type {
-  OutputElement, TextElement, ImageElement, AudioElement,
-  VideoElement, CardElement, FileElement, CardField, CardButton,
-} from '@zhin.js/ai';
-
-export {
-  MemoryStorageBackend,
-  DatabaseStorageBackend,
-  createSwappableBackend,
-} from '@zhin.js/ai';
-export type { StorageBackend, DbModel } from '@zhin.js/ai';
-
-export {
-  estimateTokens,
-  estimateMessagesTokens,
-  splitMessagesByTokenShare,
-  chunkMessagesByMaxTokens,
-  computeAdaptiveChunkRatio,
-  resolveContextWindowTokens,
-  evaluateContextWindowGuard,
-  summarizeWithFallback,
-  summarizeInStages,
-  pruneHistoryForContext,
-  compactSession,
-  DEFAULT_CONTEXT_TOKENS,
-} from '@zhin.js/ai';
-export type { ContextWindowSource, ContextWindowInfo, ContextWindowGuardResult, PruneResult } from '@zhin.js/ai';
-
 // ── Agent-specific modules ──
 
 export { AIService } from './service.js';
