+++ packages/im/agent/src/init/message-media.ts
@@ -1,4 +1,5 @@
-import type { ContentPart, Message, QuotedMessagePayload } from '@zhin.js/core';
+import type { Message, QuotedMessagePayload } from '@zhin.js/core'
+import type { ContentPart } from '@zhin.js/ai';
 import { Message as MessageNs } from '@zhin.js/core';
 
 /**
