+++ packages/im/agent/src/init/register-db-upgrade.ts
@@ -3,7 +3,7 @@
  */
 import './types.js';
 import { getPlugin } from '@zhin.js/core';
-import type { AIConfig } from '@zhin.js/core';
+import type { AIConfig } from '@zhin.js/ai';
 import type { AIServiceRefs } from './shared-refs.js';
 import { activateAiDatabaseStorage } from './activate-ai-database-storage.js';
 import {
