+++ packages/im/agent/src/routing/runtime-binding.ts
@@ -1,4 +1,4 @@
-import type { AIProvider } from '@zhin.js/core';
+import type { AIProvider } from '@zhin.js/ai';
 import type { ResolvedAgentBinding } from '../config/types.js';
 import type { ZhinAgentConfig } from '../zhin-agent/config.js';
