+++ packages/im/zhin/src/index.ts
@@ -7,17 +7,31 @@ export * from '@zhin.js/core';
  * Re-export @zhin.js/agent（会话、ZhinAgent、init、内置工具等）
  */
 export {
-  initAgentModule,
-  initAgentModule as initAIModule,
-  AIService,
   SessionManager,
   createMemorySessionManager,
   createDatabaseSessionManager,
-  ZhinAgent,
   createContextManager,
   ConversationMemory,
-  UserProfileStore,
   RateLimiter,
+  parseOutput,
+  renderToPlainText,
+  renderToSatori,
+  detectTone,
+  estimateTokens,
+  estimateMessagesTokens,
+  compactSession,
+  DEFAULT_CONTEXT_TOKENS,
+  filterTools,
+  ModelRegistry,
+  extractModelRoot,
+  computeTierScore
+} from '@zhin.js/ai';
+export {
+  initAgentModule,
+  initAgentModule as initAIModule,
+  AIService,
+  ZhinAgent,
+  UserProfileStore,
   SubagentManager,
   PersistentCronEngine,
   registerAIHook,
@@ -31,50 +45,30 @@ export {
   clearAgentPromptContributors,
   resolveAgentPromptSections,
   resolveAgentPromptMarkdown,
-  parseOutput,
-  renderToPlainText,
-  renderToSatori,
   loadBootstrapFiles,
   loadSoulPersona,
   loadToolsGuide,
-  loadAgentsMemory,
-  detectTone,
-  estimateTokens,
-  estimateMessagesTokens,
-  compactSession,
-  DEFAULT_CONTEXT_TOKENS,
+  loadAgentsMemory
 } from '@zhin.js/agent';
 
 export type {
-  ZhinAgentConfig,
-  OnChunkCallback,
+  AIConfig,
+  AIProvider,
+  AgentTool,
+  ChatMessage,
+  ContentPart,
+  ProviderConfig,
+  ProviderCapabilities,
   ContextManager,
   ContextConfig,
   MessageRecord,
   ConversationMemoryConfig,
   RateLimitConfig,
   RateLimitResult,
-  SubagentOrigin,
-  SubagentResultSender,
-  SpawnOptions,
-  SubagentManagerOptions,
-  CronJobRecord,
-  CronRunner,
-  AddCronFn,
-  PersistentCronEngineOptions,
-  CronManager,
   ContextWindowSource,
   ContextWindowInfo,
   ContextWindowGuardResult,
   PruneResult,
-  BootstrapFile,
-  ContextFile,
-  MessageReceivedEvent,
-  MessageSentEvent,
-  SessionCompactEvent,
-  SessionNewEvent,
-  AgentBootstrapEvent,
-  ToolCallEvent,
   Tone,
   OutputElement,
   TextElement,
@@ -84,8 +78,29 @@ export type {
   CardElement,
   FileElement,
   CardField,
-  CardButton,
-  RichSystemPromptContext,
+  CardButton
+} from '@zhin.js/ai';
+export type {
+  ZhinAgentConfig,
+  OnChunkCallback,
+  SubagentOrigin,
+  SubagentResultSender,
+  SpawnOptions,
+  SubagentManagerOptions,
+  CronJobRecord,
+  CronRunner,
+  AddCronFn,
+  PersistentCronEngineOptions,
+  CronManager,
+  BootstrapFile,
+  ContextFile,
+  MessageReceivedEvent,
+  MessageSentEvent,
+  SessionCompactEvent,
+  SessionNewEvent,
+  AgentBootstrapEvent,
+  ToolCallEvent,
+  RichSystemPromptContext
 } from '@zhin.js/agent';
 
 /**
